Add unique and total word counts to WordCount.java
Finished working on task 1. Program now successfully outputs
the number of unique words and the number of total words in
addition to the number of instances of each word.

I also cleaned up the file structure of the task-1 folder,
and re-wrote my Makefile since the old one was not building
the program properly.

Lastly, I lengthened the input file to better test funcitonality.