Implement Node structure to handle server interactions
The Node structure provides easy methods to interact with
the server methods from the main function. In order to
call server methods from the server side, simply use
self.addr.<method>. This can be used to pass data to
and from the server after it has been spawned.

All server information such as votes and state will
still be handled by the server, but the vector of
clients and serve handle are contained within the node.

It functions by spawning a client for the server itself.
I have not tested with multiple connections yet, but
that is my next step.