Implement logging and begin leader election framework
This change adds the log and env_logger crates to allow logging
functionality. It also creates log and log_staging vectors of
tuples to track past operations and current log value.

Also in this change is the beginnings of the leader election
system. We still need a way to pass server identification
over rpc to keep track of who is the leader. This will likely
require changes to the clients vector in the node, perhaps
making it a node of tuples or structs such that a unique id
may be associated with each client.